requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
supabase==2.0.0
groq==0.4.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
Pillow==10.2.0
aiofiles==23.2.1
jinja2==3.1.3
httpx==0.24.1

# Authentication
werkzeug>=3.0.0
python-jose[cryptography]==3.3.0
email-validator==2.3.0

# Scanner
paddlepaddle==3.0.0
paddleocr>=2.10.0
opencv-python>=4.9.0
numpy<2.0.0
opencv-python